+++ rhine-examples/src/EventClock.hs
@@ -31,20 +31,20 @@ type EventIO = EventChanT String IO
 -- ** Sample data
 
 -- | Output "Hello World!" every second.
-message :: Monad m => SyncSF m (HoistClock IO EventIO (Millisecond 1000)) () String
+message :: Monad m => ClSF m (HoistClock IO EventIO (Millisecond 1000)) () String
 message = arr $ const "Hello World!"
 
 -- | Perform a random computation, using a lot of CPU time.
-randomNumbers :: MonadIO m => Behaviour m td String
-randomNumbers = arrMSync_ $ liftIO $ do
+randomNumbers :: MonadIO m => Behaviour m time String
+randomNumbers = constMCl $ liftIO $ do
   m <- randomRIO (4, 6 :: Integer)
   x <- randomRIO (-3, 3 :: Double)
   return $ "Random number: " ++ printf "%.3f" (iterate tan x !! (10 ^ m))
 
 -- | Each time an event arrives, this function is called.
 --   It simply outputs the event on the console.
-handleEvents :: (MonadIO m, Tag cl ~ String) => SyncSF m cl () ()
-handleEvents = theTag >>> arrMSync (putStrLn >>> liftIO)
+handleEvents :: (MonadIO m, Tag cl ~ String) => ClSF m cl () ()
+handleEvents = tagS >>> arrMCl (putStrLn >>> liftIO)
 
 -- * Running the subsystems in the same thread, or in separate threads
 
@@ -70,10 +70,10 @@ threadsExample = do
 --   that tries to run in parallel at a higher frequency,
 --   but will be caused to lag because it is run in the same thread
 --   as a computationally expensive one.
-responsive :: SyncSF IO (Millisecond 100) () ()
+responsive :: ClSF IO (Millisecond 100) () ()
 responsive = timeInfo >>> proc TimeInfo {..} -> do
-  arrMSync putStrLn -< "Current time: " ++ show sinceStart
-  arrMSync putStrLn -< "Real time " ++ (if tag then "" else "UN") ++ "successful"
+  arrMCl putStrLn -< "Current time: " ++ show sinceInit
+  arrMCl putStrLn -< "Real time " ++ (if tag then "" else "UN") ++ "successful"
 
 
 -- | Producing random numbers using a lot of CPU causes all subsystems on the same thread to lag.
@@ -84,7 +84,7 @@ randomsExample = runEventChanT $ flow wholeSystem
     emitEventSystem   = randomNumbers >-> emitS @@ busy
     handleEventSystem = handleEvents            @@ EventClock
     eventSystem = emitEventSystem **@ concurrentlyWithEvents @** handleEventSystem
-    responsiveSystem = liftSyncSFAndClock responsive @@ liftClock waitClock -- TODO This can be lifted in one go
+    responsiveSystem = liftClSFAndClock responsive @@ liftClock waitClock -- TODO This can be lifted in one go
     wholeSystem = eventSystem **@ concurrentlyWithEvents @** responsiveSystem