+++ rhine/src/FRP/Rhine.hs
@@ -9,7 +9,7 @@ import FRP.Rhine
 import FRP.Rhine.Clock.Realtime.Millisecond
 
 main :: IO ()
-main = flow $ arrMSync_ (putStrLn "Hello World!") @@ (waitClock :: Millisecond 100)
+main = flow $ constMCl (putStrLn "Hello World!") @@ (waitClock :: Millisecond 100)
 @
 -}
 module FRP.Rhine (module X) where
@@ -22,6 +22,6 @@ import FRP.Rhine.Clock            as X
 import FRP.Rhine.Reactimation     as X
 import FRP.Rhine.ResamplingBuffer as X
 import FRP.Rhine.Schedule         as X
-import FRP.Rhine.SF               as X
-import FRP.Rhine.SF.Combinators   as X
-import FRP.Rhine.SyncSF           as X
+import FRP.Rhine.SN               as X
+import FRP.Rhine.SN.Combinators   as X
+import FRP.Rhine.ClSF           as X