+++ rhine/src/FRP/Rhine/Clock/Realtime/Millisecond.hs
@@ -17,7 +17,7 @@ import Data.Vector.Sized (Vector, fromList)
 
 -- rhine
 import FRP.Rhine
-import FRP.Rhine.Clock.Step
+import FRP.Rhine.Clock.FixedStep
 import FRP.Rhine.ResamplingBuffer.Collect
 import FRP.Rhine.ResamplingBuffer.Util
 
@@ -32,13 +32,13 @@ The tag of this clock is 'Bool',
 where 'True' represents successful realtime,
 and 'False' a lag.
 -}
-newtype Millisecond (n :: Nat) = Millisecond (RescaledClockS IO (Step n) UTCTime Bool)
+newtype Millisecond (n :: Nat) = Millisecond (RescaledClockS IO (FixedStep n) UTCTime Bool)
 -- TODO Consider changing the tag to Maybe Double
 
 instance Clock IO (Millisecond n) where
-  type TimeDomainOf (Millisecond n) = UTCTime
-  type Tag          (Millisecond n) = Bool
-  startClock (Millisecond cl) = startClock cl
+  type Time (Millisecond n) = UTCTime
+  type Tag  (Millisecond n) = Bool
+  initClock (Millisecond cl) = initClock cl
 
 
 -- | This clock simply sleeps 'n' milliseconds after each tick.
@@ -46,9 +46,9 @@ instance Clock IO (Millisecond n) where
 --   Consequently, the tag is constantly 'False',
 --   since the clock will accumulate the computation time as lag.
 sleepClock :: KnownNat n => Millisecond n
-sleepClock = sleepClock_ Step
+sleepClock = sleepClock_ FixedStep
   where
-    sleepClock_ :: Step n -> Millisecond n
+    sleepClock_ :: FixedStep n -> Millisecond n
     sleepClock_ cl = Millisecond $ RescaledClockS cl $ const $ do
       now <- getCurrentTime
       let ticks = arrM_ $ do
@@ -67,7 +67,7 @@ sleepClock = sleepClock_ Step
 --   If the next tick should already have occurred,
 --   the tag is set to 'False', representing a failed real time attempt.
 waitClock :: KnownNat n => Millisecond n
-waitClock = Millisecond $ RescaledClockS Step $ \_ -> do
+waitClock = Millisecond $ RescaledClockS FixedStep $ \_ -> do
   initTime <- getCurrentTime
   let
     runningClock = arrM $ \(n, ()) -> do
@@ -82,7 +82,7 @@ waitClock = Millisecond $ RescaledClockS Step $ \_ -> do
   return (runningClock, initTime)
 
 
--- TODO It would be great if this could be directly implemented in terms of downsampleStep
+-- TODO It would be great if this could be directly implemented in terms of downsampleFixedStep
 downsampleMillisecond
   :: (KnownNat n, Monad m)
   => ResamplingBuffer m (Millisecond k) (Millisecond (n * k)) a (Vector n a)
@@ -96,7 +96,7 @@ downsampleMillisecond = collect >>-^ arr (fromList >>> assumeSize)
 
 -- | Two 'Millisecond' clocks can always be scheduled deterministically.
 scheduleMillisecond :: Schedule IO (Millisecond n1) (Millisecond n2)
-scheduleMillisecond = Schedule startSchedule'
+scheduleMillisecond = Schedule initSchedule'
   where
-    startSchedule' (Millisecond cl1) (Millisecond cl2)
-      = startSchedule (rescaledScheduleS scheduleStep) cl1 cl2
+    initSchedule' (Millisecond cl1) (Millisecond cl2)
+      = initSchedule (rescaledScheduleS scheduleFixedStep) cl1 cl2