+++ rhine/src/FRP/Rhine/ResamplingBuffer/Interpolation.hs
@@ -15,16 +15,16 @@ import FRP.Rhine.ResamplingBuffer.Util
 linear
   :: ( Monad m, Clock m cl1, Clock m cl2
      , VectorSpace v
-     , Groundfield v ~ Diff (TimeDomainOf cl1)
-     , Groundfield v ~ Diff (TimeDomainOf cl2)
+     , Groundfield v ~ Diff (Time cl1)
+     , Groundfield v ~ Diff (Time cl2)
      )
   => v -- ^ The initial velocity (derivative of the signal)
   -> v -- ^ The initial position
   -> ResamplingBuffer m cl1 cl2 v v
 linear initVelocity initPosition
-  =    (derivativeFrom initPosition &&& syncId) &&& timeInfoOf sinceStart
+  =    (derivativeFrom initPosition &&& clId) &&& timeInfoOf sinceInit
   ^->> keepLast ((initVelocity, initPosition), 0)
-  >>-^ proc ((velocity, lastPosition), sinceStart1) -> do
-    sinceStart2 <- timeInfoOf sinceStart -< ()
-    let diff = sinceStart2 - sinceStart1
+  >>-^ proc ((velocity, lastPosition), sinceInit1) -> do
+    sinceInit2 <- timeInfoOf sinceInit -< ()
+    let diff = sinceInit2 - sinceInit1
     returnA -< lastPosition ^+^ velocity ^* diff