+++ rhine/ChangeLog.md
@@ -29,7 +29,7 @@ every major version in `dunai` triggers a major version in `rhine`.
 
 * Added `FRP.Rhine.Clock.Realtime.Stdin` (console keyboard event clock)
 * Added `FRP.Rhine.Clock.Select` (event selection clock)
-* Added `FRP.Rhine.SyncSF.Except` (synchronous exception handling)
+* Added `FRP.Rhine.ClSF.Except` (synchronous exception handling)
 
 ## 0.1.0.0